+++ packages/forge-config/lib/index.d.ts
@@ -1,0 +1,22 @@
+// Generate by [js2dts@0.3.3](https://github.com/whxaxes/js2dts#readme)
+
+/**
+ * Parse config from a file
+ *
+ * @public
+ * @static
+ * @param {string} configPath
+ * @returns object
+ * @throws Error when file not found
+ * @example
+ const { parseConfig } = require('@arcblock/forge-config');
+ const config = parseConfig('./forge.toml');
+ */
+declare function parse(configPath: string): any;
+declare const _Lib: _Lib.T100;
+declare namespace _Lib {
+  export interface T100 {
+    parse: typeof parse;
+  }
+}
+export = _Lib;